requirements.txt
numpy==1.24.2
opencv_contrib_python==4.5.5.62
opencv_python==4.6.0.66
opencv_python_headless==4.5.4.60
Pillow==9.2.0
Pillow==10.0.0
protobuf==3.20.3
streamlit==1.21.0
streamlit_extras==0.2.7
streamlit-option-menu==0.3.6
torch==1.13.1
google-cloud-vision
